// src/lib/ride-backend-types.ts
export type TripStatus =
    | 'pending'
    | 'searching_driver'
    | 'offered'
    | 'accepted'
    | 'driver_on_the_way'
    | 'driver_arrived'
    | 'trip_started'
    | 'completed'
    | 'cancelled';